Setting up the main-storyboard
I made changes to the project.  I will not be using landscape in this
application as it will be too complicated for the users.  Keeping it
only portrait will benefit the user the most.  I also added all 5
screens that will be in the application, even though for this class I
may not be doing all screens.  This took me about 50 minutes to put
together as We did not use tab bar controllers as navigation before.
Now I will create the remaining icons for the tab bar and then attach
the map so that I can run the application and see how it looks.  At
this time I cannot run app because the map view is not connected yet.