Log-In Auth Completed the right way
I had a bug in the Facebook log in, so I finally was able to fix it and
now I get the Facebook credentials that I need to move forward.  This
took me a while but now I can start on the Favorites page.  I also was
able to do more design UI stuff so that will be new.